cleanup

+++ test/Market.t.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.0;
 
 import "forge-std/Test.sol";
 import "forge-std/console.sol";
-import "../src/gnosispm/Market.sol";
-import "../src/gnosispm/MarketMaker.sol";
-import "../src/gnosispm/Event.sol";
-import "../src/gnosispm/MockERC20.sol"; // MockERC20 is a mock contract for ERC20
+import "../src/prediction-market/Market.sol";
+import "../src/prediction-market/MarketMaker.sol";
+import "../src/prediction-market/Event.sol";
+import "../src/prediction-market/MockERC20.sol"; // MockERC20 is a mock contract for ERC20
 
 contract MarketTest is Test {
     Market public market;
@@ -17,7 +17,7 @@ contract MarketTest is Test {
     address public creator = address(1);
     address public buyer = address(2);
     address public seller = address(3);
-    uint public fundingAmount = 1000;
+    uint256 public fundingAmount = 1000;
 
     function setUp() public {
         // Deploy the mock ERC20 token
@@ -26,22 +26,9 @@ contract MarketTest is Test {
         collateralToken.mint(buyer, 10 ** 18);
         collateralToken.mint(seller, 10 ** 18);
 
-        // Deploy the mock Event contract
-        // eventContract = new MockEvent(address(collateralToken));
         eventContract = new CategoricalEvent(collateralToken, 2);
-
         marketMaker = new MarketMaker();
-
-        // Deploy the Market contract
         market = new Market(creator, eventContract, marketMaker, 10000);
-
-        // Initialize the Market contract
-        // market.initialize(address(eventContract), creator);
-
-        // Set initial state for the Market contract
-        // market.setCreator(creator);
-        // market.setStage(Market.Stages.MarketCreated);
-        // market.setFee(10000); // 1% fee
     }
 
     function testFundMarket() public {
@@ -52,7 +39,7 @@ contract MarketTest is Test {
         vm.stopPrank();
 
         assertEq(market.funding(), fundingAmount);
-        assertEq(uint(market.stage()), uint(MarketData.Stages.MarketFunded));
+        assertEq(uint256(market.stage()), uint256(MarketData.Stages.MarketFunded));
     }
 
     function testCloseMarket() public {
@@ -62,14 +49,14 @@ contract MarketTest is Test {
         market.close();
         vm.stopPrank();
 
-        assertEq(uint(market.stage()), uint(MarketData.Stages.MarketClosed));
+        assertEq(uint256(market.stage()), uint256(MarketData.Stages.MarketClosed));
     }
 
     function testWithdrawFees() public {
         vm.startPrank(creator);
         collateralToken.approve(address(market), fundingAmount);
         market.fund(fundingAmount);
-        uint fees = market.withdrawFees();
+        uint256 fees = market.withdrawFees();
         vm.stopPrank();
 
         // No fees?
@@ -84,7 +71,7 @@ contract MarketTest is Test {
 
         vm.startPrank(buyer);
         collateralToken.approve(address(market), fundingAmount);
-        uint cost = market.buy(0, 10, 100);
+        uint256 cost = market.buy(0, 10, 100);
         vm.stopPrank();
 
         // TODO - is this right?
@@ -99,12 +86,12 @@ contract MarketTest is Test {
 
         vm.startPrank(buyer);
         collateralToken.approve(address(market), fundingAmount * 2);
-        uint cost = market.buy(0, 10, 100);
+        uint256 cost = market.buy(0, 10, 100);
 
         // FAILS HERE!
         // tradeImpl tries to transfer 115792089237316195423570985008687907853269984665640564039457584007913129639934
         collateralToken.approve(address(market), fundingAmount);
-        uint profit = market.sell(0, 5, 1);
+        uint256 profit = market.sell(0, 5, 1);
         vm.stopPrank();
         assertGt(profit, 0);
     }
@@ -118,7 +105,7 @@ contract MarketTest is Test {
         vm.startPrank(seller);
         // tradeImpl tries to transfer 115792089237316195423570985008687907853269984665640564039457584007913129639934
         collateralToken.approve(address(market), fundingAmount);
-        uint profit = market.shortSell(0, 10, 1);
+        uint256 profit = market.shortSell(0, 10, 1);
         vm.stopPrank();
         assertGt(profit, 0);
     }
